requirements.txt
mock
nose
pexpect
pyhamcrest
